Move snapshottable parent mixin into separete file
Move the snapshottable parent (the object around which the scope
is created) into separate file
